Update README.md
This repository contains the primary code used for the single-cell data analysis presented in the study titled "STAT3-Orchestrated Gene Expression Signatures in Esophageal Squamous Cell Carcinoma Uncovered by Single-Cell Sequencing." The code facilitates comprehensive analyses, including the identification and classification of cell types, the calculation and visualization of cell-specific markers, the detection and presentation of differentially expressed genes, the construction and analysis of pseudotime trajectories to investigate cell fate decisions, and the assessment of gene expression patterns related to cellular processes. These tools are designed to ensure reproducibility and to support further exploration of single-cell transcriptomics in cancer research.